+++ src/Plugins/Litedown/Parser/LinkAttributesSetter.js
@@ -15,6 +15,10 @@ function setLinkAttributes(tag, linkInfo, attrName)
 		title = url.substr(pos).replace(/^\s*\S/, '').replace(/\S\s*$/, '');
 		url   = url.substr(0, pos);
 	}
+	if (/^<.+>$/.test(url))
+	{
+		url = url.replace(/^<(.+)>$/, '$1').replace(/\\>/g, '>');
+	}
 
 	tag.setAttribute(attrName, decode(url));
 	if (title > '')