+++ src/Plugins/Litedown/Parser/Passes/Links.php
@@ -22,6 +22,10 @@ class Links extends AbstractPass
 		{
 			$this->parseInlineLinks();
 		}
+		if ($this->text->indexOf('<') !== false)
+		{
+			$this->parseAutomaticLinks();
+		}
 		if ($this->text->hasReferences)
 		{
 			$this->parseReferenceLinks();
@@ -72,6 +76,34 @@ class Links extends AbstractPass
 		}
 
 		return $labels;
+	}
+
+	/**
+	* Parse automatic links markup
+	*
+	* @return void
+	*/
+	protected function parseAutomaticLinks()
+	{
+		preg_match_all(
+			'/<[-+.\\w]++([:@])[^\\x17\\s>]+?(?:>|\\x1B7)/',
+			$this->text,
+			$matches,
+			PREG_OFFSET_CAPTURE
+		);
+		foreach ($matches[0] as $i => $m)
+		{
+			// Re-escape escape sequences in automatic links
+			$content  = substr($this->text->decode(str_replace("\x1B", "\\\x1B", $m[0])), 1, -1);
+			$startPos = $m[1];
+			$endPos   = $startPos + strlen($m[0]) - 1;
+
+			$tagName  = ($matches[1][$i][0] === ':') ? 'URL' : 'EMAIL';
+			$attrName = strtolower($tagName);
+
+			$this->parser->addTagPair($tagName, $startPos, 1, $endPos, 1)
+			             ->setAttribute($attrName, $content);
+		}
 	}
 
 	/**